Extract rendering methods
